moved settings.dart back to base /lib
